Factor out witness generation
