+++ noir-r1cs/src/bin/prover.rs
@@ -12,7 +12,6 @@ use {
         utils::*,
         whir_utils::*,
     },
-    rand::Rng,
     spongefish::{
         codecs::arkworks_algebra::{FieldToUnitDeserialize, FieldToUnitSerialize, UnitToField},
         DomainSeparator, ProverState, VerifierState,
